+++ tests/testthat/test_print_Appender.R
@@ -1,11 +1,11 @@
 context("print_Appender")
 
 
+
+
 test_that("print_Appender works as expected", {
   tf <- tempfile()
   on.exit(unlink(tf))
-
-
 
   Appender$new()
   AppenderConsole$new()
@@ -25,29 +25,4 @@ test_that("print_Appender works as expected", {
   AppenderFileRotatingTime$new(tf, layout = LayoutJson$new(), age = "2 years", size = 9000)
   AppenderFile$new(tf)
 
-
 })
-
-
-
-<AppenderConsole> [all]
-  Destination:  console
-  LayoutFormat:  %L [%t] %m %f
-
-
-
-<AppenderFile> [all]
-  Destination:  "/tmp/RtmpWYA891/file5ac1dc5db5d"
-  LayoutFormat:  %L [%t] %m %f
-
-
-<AppenderJson> [all]
-  Destination:  "/tmp/RtmpWYA891/file5ac1dc5db5d"
-  LayoutJson:  jsonlite::toJSON
-
-
-
-<AppenderFileRotating> [all]
-  Destination:  "/tmp/RtmpWYA891/file5ac1dc5db5d"
-  LayoutJson:  jsonlite::toJSON
-  Age: "2 Years"